+++ great_expectations/self_check/util.py
@@ -2006,10 +2006,15 @@ def _create_bigquery_engine() -> Engine:
     gcp_project = os.getenv("GE_TEST_BIGQUERY_PROJECT")
     if not gcp_project:
         raise ValueError(
-            "Environment Variable GE_TEST_BIGQUERY_PROJECT is required to run expectation tests"
+            "Environment Variable GE_TEST_BIGQUERY_PROJECT is required to run BigQuery expectation tests"
         )
     return create_engine(f"bigquery://{gcp_project}/{_bigquery_dataset()}")
 
 
 def _bigquery_dataset() -> str:
-    return os.getenv("GE_TEST_BIGQUERY_DATASET")
+    dataset = os.getenv("GE_TEST_BIGQUERY_DATASET")
+    if not dataset:
+        raise ValueError(
+            "Environment Variable GE_TEST_BIGQUERY_DATASET is required to run BigQuery expectation tests"
+        )
+    return dataset